Backend/src/codeedex_project/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
Pillow